LPD-67397 baseline
# breaking

## What modules/apps/export-import/export-import-api/src/main/java/com/liferay/exportimport/vulcan/batch/engine/ExportImportVulcanBatchEngineTaskItemDelegate.java

New methods:
    getResourceClassName
    isApplicableExternalReferenceCode

Renamed methods:
    getItemClassName -> getModelClassName
    getItemModelName -> getModelName
    getLabel -> getLabelLanguageKey

## Why

getModelClassName: it is the internal className. Used to retrieve the systemEventDeletions and to save the ExportImportReportEntry
getModelName: removed the 'item' from the name for coherence
getLabelLanguageKey: it is actually a languageKey that gets translated
getResourceClassName: unique class name. Hardcoded to the resourceImpl className. Used for controls and count's StagedModelType
isApplicableExternalReferenceCode: a new method to filter the actual externalReferenceCodes that correspond to each resource

----